[frontend] add validation work to draft overview in read only mode
